requirements.txt
# Core dependencies
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# CLI and output formatting
rich>=13.7.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.12.0
mypy>=1.7.0
